implement json test in jdbi-jackson, get null test green
